MLBEDSW-11736 Fix performance estimator overflow
Upgraded some calculations/calls to use int64 to avoid overflow.

+++ ethosu/regor/architecture/architecture.hpp
@@ -236,11 +236,11 @@ enum OpScheduling
 /// </summary>
 struct ElementAccess
 {
-    int ifmRead[2] = {0, 0};
-    int ofmWrite = 0;
+    int64_t ifmRead[2] = {0, 0};
+    int64_t ofmWrite = 0;
     int weightsRefetch = 0;
     int constRead[2] = {0, 0};
-    int tmpRead = 0, tmpWrite = 0;
+    int64_t tmpRead = 0, tmpWrite = 0;
 };
 
 /// <summary>
